Tools decomposition and refactoring from core to root
