+++ sgr_deep_research/core/agents/base_agent.py
@@ -13,7 +13,7 @@ from openai.types.chat import ChatCompletionFunctionToolParam
 from sgr_deep_research.core.models import AgentStatesEnum, ResearchContext
 from sgr_deep_research.core.prompts import PromptLoader
 from sgr_deep_research.core.stream import OpenAIStreamingGenerator
-from sgr_deep_research.core.tools import (
+from sgr_deep_research.tools import (
     # Base
     BaseTool,
     ClarificationTool,
