+++ sgr_deep_research/core/agents/sgr_agent.py
@@ -3,7 +3,7 @@ import uuid
 from typing import Type
 
 from sgr_deep_research.core.agents.base_agent import BaseAgent
-from sgr_deep_research.core.tools import (
+from sgr_deep_research.tools import (
     AgentCompletionTool,
     BaseTool,
     ClarificationTool,
