+++ sgr_deep_research/core/agents/sgr_so_tools_agent.py
@@ -3,7 +3,7 @@ import uuid
 from typing import Type
 
 from sgr_deep_research.core.agents.sgr_tools_agent import SGRToolCallingResearchAgent
-from sgr_deep_research.core.tools import BaseTool, ReasoningTool
+from sgr_deep_research.tools import BaseTool, ReasoningTool
 from sgr_deep_research.settings import get_config
 
 logging.basicConfig(
